pyproject.toml
[project]
name = "tasks2bookstack"
version = "0.1.0"
description = "CalDAV Tasks 2 Bookstack Sync"
readme = "README.md"
authors = [
    { name = "Anchit Bajaj", email = "ab@abifog.com" }
]
requires-python = ">=3.12"
dependencies = [
    "caldav>=2.0.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[project.scripts]
tasks2bookstack = "tasks2bookstack.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.pyright]
typeCheckingMode = "standard"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "types-pyyaml>=6.0.12.20250822",
]
